Finishes off the browser tests
This took forever because I was trying to solve a more in depth problem.
The browser controller is basically just a pass through and that bothers me.
A lot of extra lines of code plus a lot of extra tests that just test pass through.

Worked on building something that could make this more of a delegation process.
Gabo and I came up with some good ideas, but nothing that feels correct yet.