Adds an express server with proxy abilities
Solves the problem of the client needing to know the address of the API
server. With this approach, the client calls back to the initiating server
and then calls are proxied from there.

I could use something like nginx to front the API server and also serve
the Angular app. At the moment, this seemed like one more thing for people
to configure, but maybe that would be a better option moving forward.