Dropping webpack dev server for now
I wasn't able to hit it from my phone (outside my dev machine)
and honestly it isn't bringing a ton to the table right now for me.

When I am in the mood to play with it more and read some docs I will
circle back, but for now I want to develop actual code.